Implement a primitive distinguisher
There's more to do. Specialization for the array and table types.
Furthermore, it's important to get an algorithm going. I'm not sure
how to go about putting stuff into the table, but from what I can tell
we can simply hold a vector of tables. The last table in the vector
is the currently active one (the one we're parsing into).

I'm not entirely sure how to do this in Rust without RefCell or
recursion.

	State: [ top ]
	Input: "{ table"
	=>
	State: [ top, table ]

A possibility is to avoid adding the element until it is popped from
the stack. This means that "table" itself will be inserted into top
once it's popped. The type looks like:

	let tree: Tree;
	let stack: Vec<RefCell<Value>>;

The easiest way of doing things is by recursion. It's also what
mathematicians and Haskell programmers love, so I'll try that.